feat(#27): separando component de analista
o component talhaolist voltou a ser como era na branch da sprint2
